backend/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
watchdog==3.0.0
pandas==2.1.4
openpyxl==3.1.2
numpy==1.24.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
